intg 1575 refactoring
